refactor(patreon): move files
